requirements.txt
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-bootstrap-datepicker-plus
django-bootstrap4
django-crispy-forms
django-embed-video
django-filter
django-flatpickr
idna
oauthlib
pycparser
pydantic
PyJWT
python3-openid
requests
requests-oauthlib
soupsieve
sqlparse
typing_extensions
tzdata
urllib3
django-flatpickr
rest_framework_simplejwt
django-widget-tweaks